requirements.txt
-i https://pypi.tuna.tsinghua.edu.cn/simple

anyio==3.6.2

certifi==2022.12.7

charset-normalizer==3.0.1

click==8.1.3

colorama==0.4.6

fastapi==0.92.0

greenlet==2.0.2

h11==0.14.0

httptools==0.5.0

idna==3.4

install==1.3.5

loguru==0.6.0

pip==23.0.1

pydantic==1.10.5

pymysql==1.0.2

pysnowflake==0.1.3

python-dotenv==0.21.1

pyyaml==6.0

requests==2.28.2

six==1.16.0

sniffio==1.3.0

sqlalchemy==2.0.4

starlette==0.25.0

tornado==6.2

typing-extensions==4.5.0

urllib3==1.26.14

uvicorn[standard]==0.20.0

watchfiles==0.18.1

websockets==10.4

win32-setctime==1.1.0

